Move StatsManager instantiation from Stats to StatsInternal.
